Separate files for inspector's 3D Transform input and PortValue.padding views
